+++ cfdna/Plot/__init__.py
@@ -1,7 +1,19 @@
 from .plot_plt import *
-
+from .plot_bokeh import *
+from . import plot_plt
+from . import plot_bokeh
 
 __doc__="""
 Plotting API
 ============
+.. currentmodule:: cfdna
+
+Generic
+-------
+.. autosummary::
+   :toctree: .
+
+   plot_plt
+   plot_bokeh
+   
 """