+++ cfdna/__init__.py
@@ -1,15 +1,15 @@
 from __future__ import absolute_import
-from .core.core import cfDNA
-from . import plot as pl
+from .core import cfDNA
+from . import plot
 from . import io
 from . import commandline as cmd
-from . import tools as tl
+from . import tools
 
 from ngsfragments import Fragments
 
 
 # This is extracted automatically by the top-level setup.py.
-__version__ = '2.0.7'
+__version__ = '2.1.2'
 
 __author__ = "Kyle S. Smith"
 
@@ -33,17 +33,30 @@ Plotting functions
 .. autosummary::
    :toctree: .
    
-   pl.plot_plt
-   pl.plot_bokeh
+   plot.plot_plt
 
 
 IO functions
-------------------
+------------
     
 .. autosummary::
    :toctree: .
    
-   io.write
-   io.read
+   io.write_h5
+   io.write_anno
+   io.write_seg
+   io.write_anno_seg
+   io.read_sam
+   io.read_h5
+
+
+Tools
+-----
+    
+.. autosummary::
+   :toctree: .
+   
+   tools.nucleosome
+   tools.cnv
 
 """