+++ doc/release-notes.rst
@@ -5,6 +5,29 @@ Release Notes
 .. role:: smaller
 
 
+Version 2.1.2
+-------------
+General updates for pip installation
+	- See installation
+
+
+Version 2.1.1
+-------------
+Function simplifications
+
+
+Version 2.1.0
+-------------
+Commandline interface created
+	- See tutorials
+
+
+Version 2.0.0
+-------------
+:class:`~cfdna.cfDNA` class overhaul
+	- Now a subclass of the ProjectFrame object
+
+
 Version 0.1.0
 -------------